Rename: Rename the dir `pixverese` to `pixverse_ai`
Rename the dir 'pixverse' to 'pixverse' because it contains a module
named as 'pixverse' and when main.py import it then the intelligence assume
that 'pixverse' dir is imported and due to that intelligence doen't work
as expected. By the way, there is no issue in working because in real
imported file is 'pixverse.py' but intelligence doesn't work.

Signed-off-by: @surajgirioffl